:recycle:  Major refactor
