+++ drbert/__init__.py
@@ -1,4 +1,4 @@
 # Importable from root
 from .data.dataset_readers import *
-from .heads import *
 from .model import DrBERT
+from .modules import *
